Speed up alignment by approx. 10%.
When we want to compute a cell of the DP matrix and find a match between the
two characters of the input strings, we know that this is already the best we
can do. We now skip the logic that checks whether an insertion or deletion is
even better.

Total runtime (on my machine) decreases by ~5% when searching for a single
adapter.