replace ReadFilter with a ProcessedReadWriter
All the code that sends a read to a different output file depending on
different criteria (too long, too short, trimmed, not trimmed) is now in
this class. This makes it a lot easier to understand where there the reads
actually go.

Also, get rid of the 'twoheaders' handling in many places. Make twoheaders
an attribute of a Sequence instead, and let the Sequence deal with it.