Implement support for demultiplexing.
To use this, include the string "{name}" in the name of the output
file. It is then interpreted as a template and a trimmed read is written
to a file in which {name} is replaced with the name of the adapter
that was found in the read.

Only single-end reads are supported right now.

Also includes some refactoring of too short/long filtering.