Finally, support for parallel execution (!)
Most of the refactoring over the last weeks was done to make this possible.
This commit finally pulls everything together and introduces a new
ParallelPipelineRunner class that can be used to run a Pipeline in parallel.

Currently, there are some limitations, but those should be easy to fix now
that the structure is in place.

The biggest limitations are that the statistics are empty and that only a
single input and output file are supported - this will be changed soon.

Also, all of this is available only on Python 3! My time is quite limited
nowadays, so I’m no longer willing to maintain Python 2 compatibility if it
takes to much work. To be clear: Cutadapt runs on Python 2, but not in
parallel.