Move package directory to src/
This goes against the currently officially recommended package layout, but
there are quite a few reasons for doing it this way. In particular, having
the package directory *not* directly importable ensures that the tests
actually run against the packaged version of the code, not against what
happens to be in the working copy.

Blog articles:

* https://hynek.me/articles/testing-packaging/
* https://blog.ionelmc.ro/2014/05/25/python-packaging/#the-structure

Recommendation in the pytest documentation:

* https://docs.pytest.org/en/latest/goodpractices.html#tests-outside-application-code