Do not fallback to single-core mode, but error out
If multi-core operation is requested (-j N with N > 1), but not supported
for the given command-line parameters, then no longer automatically fall
back to single-core mode. Instead, exit with an error. Also, if -j is
used on Python 2, then we exit with an error.

This should give users less of a suprise when they find the option and start
to use it.

When no -j is given, the behavior is still to use as many cores as we can.