Avoid infinite loop when using --no-trim and --info-file
Sequence and Match objects are arranged in a kind of linked list:

A (trimmed) sequence has a .match attribute (of type Match) that contains
information about the adapter and where the match was found. The Match
object has a .read attribute that in turn points to the Sequence object
representing how the read was before trimming. The routine that prints out
the info file traverses this linked list in order to find out what the
original read was. The problem is that setting --no-trim turned the list
into a cycle. Creating a copy of the original read should resolve this.

This closes #296