Remove Sequence object’s "match" attribute
The .match attribute of a Sequence was used to store the most recent adapter
match. This was an optimization, but it is not measurably slower to just
pass a the list of matches into all filters and modifiers. It is not as
concise, but it makes it more explicit which data flows where.

This is a preparation for the switch to dnaio, where Sequence objects
don’t have a match attribute (it wouldn’t make sense).