Add a "remove" parameter to Adapter
It allows to set which part of the read should be removed: a prefix or
suffix of the read. Previously, this was entwined with the adapter type
specified via the 'where' parameter.

Now, there is a clear distinction:

'where' specifies where the adapter sequence is allowed to occur and
'remove' describes what to remove if it was found.