When only R1 adapters given, use --pair-filter=both for untrimmed output
When only R1 or only R2 adapters are given, the `--pair-filter` setting
is now forced to `both` for the `--discard-untrimmed` (and
`--untrimmed-(paired-)output`) filters.

Otherwise, with the default of `--pair-filter=any`, all pairs would be
considered untrimmed because one of the reads in the pair is always
untrimmed.

Close #257