Speed up trimming of many prefix or suffix adapters
The following conditions need to be met:

- multiple prefix adapters (anchored 5') are given or
  multiple suffix adapters (anchored 3') are given
- the --no-indel switch is used
- the error rate allows at most 2 errors
- no wildcards are used (neither in adapter nor the read)

If all of the above is true, then an index is built for all possible
prefixes or suffixes that could match one of the adapters, and instead of
checking each adapter separately, lookup in the index is used to
determine whether the adapter occurs.

This speeds up demultiplexing quite a lot (factor six in one of my tests on
a dataset with 96 barcodes).