Circumvent soft limit on the number of open files when demultiplexing
When opening a file during demultiplexing fails because of too many open
files, we simply raise the soft limit and re-try. This should reduce the
number of times the user sees a "Too many open files" error, but it does not
help of course if the hard limit is too low.

See #320