With --cores=1, no longer use pigz subprocesses for output
This introduces a FileOpener class, which is used in place of xopen and
dnaio.open for opening files. It can be configured to enable or disable
usage of subprocesses.

Currently, a pigz process is still spawned for gzip *input* files as pigz is
more efficient at decompressing gzip (even with a single core).

See issue #290