Do not search for an exact before an approximate match
This was intended as an optimization. Since str.find is so fast, the idea
was to skip running the more expensive approximate algorithm if an exact
match can be found.

However, this only really works if a high fraction of reads actually have an
exact adapter occurrence. On my test data with 10% of reads having an exact
occurrence, it is actually faster not do the exact search. (15% of reads
have an inexact adapter occurrence in this dataset.)

In any case, the code gets simpler and if we reintroduce this, the
optimization needs to be part of the alignment routine itself (in the
Aligner).