Remove the unhelpful .endswith() and .startswith() "optimizations"
The optimization of searching for an exact match first with str.find has
already been removed in 367a380ede1ded68601a30c2f05d25b37a569f21 for the
regular 5' and 3' adapter types, but the same reasoning applies to the other
adapters types, so remove the "startswith()" and "endswith()" checks for the
anchored and non-internal 5' and 3' adapter types as well. This is actually
faster in at least one test dataset.