Fix --pair-adapters not finding some pairs
Previously, we would find the the best matching adapter in R1 and
then look for its partner in R2. If no partner was found, no
match was reported. This does not work when another adapter in R1
would actually have a partner in R2.

Now, we search for all adapter pairs one by one and take the best
overall pair (with the most matches) to avoid this problem.

This is slower, but correct.

See #570