Fix: Partial matches with N wildcards were sometimes not found correctly
When determining whether the adapter XNNNCGC matches the reference CCCTTT,
an "effective length" of the overlapping region is computed that is supposed
to not include the N wildcards. This computation was incorrect for partial
matches and can lead to a too short effective length, preventing a match
from being found.

Closes #654