Fix: Do not hang on errors in worker processes
When an error occurs in a worker process (for example, on encountering an
incorrecty formatted FASTQ record), the main process for some reason waited
for its children to finish (although the daemon flag was set to True for
them), which meant that it an exception raised in a worker makes the program
hang.

With this, we explicitly terminate all child processes before exiting, which
appears to work.

Closes #528