Do not report a match for ambiguous sequences
This only applies when an index is used.

At index creation time, ambiguous sequences (those that would match two or
more adapters equally well), are now *removed* from the index. That is, when
they are encountered in a read, they will not be assigned to any adapter.

The intention is to reduce bias because we previously would report a match
to one of the adapters (but which one was more or less randomly determined
at index creation time).