Always create an index if there are at least two suitable adapters
The previous threshold was six, but benchmarking shows that using an index
is very slightly slower for two or three adapters, and it breaks even for
about four. The slight slowdown for these numbers of adapters is worth it
because we get more consistent behavior: If there are multiple prefix or
suffix adapters, an index is now always built.

This is especially important because we want to make the feature of checking
for collisions between adapter sequences more prominent, which is only done
when indexing. We want to check for collisions for any number of adapters
greater one.