Fix unoptimized wheels
setuptools starting from version 75.7.0 changed how CFLAGS work: They no
longer are added to the CFLAGS Python was compiled with, but replace them.

With this change, our CFLAGS of `-g0` get used directly and the
wheels are compiled without proper optimizations because `-O` is missing.

Within cibuildwheel, the Python compiler options are:

    -fno-strict-overflow -Wsign-compare -DNDEBUG -g -O3 -Wall

The fix here is to just set the CFLAGS to `-O3` because that is the only
option that we really need.

Found by @rhpvorderman in https://github.com/marcelm/cutadapt/issues/843